requirements.txt
fastapi
uvicorn
pandas
pydantic
python-dotenv
requests
